Fix ReadOnlyLedgerHandle leak issue when checkAllLedgers. (#4468)
When the Auditor checkAllledgers, it will open the ledger with NoRecovery mode, it will register the listeners to the AbstractZkLedgerManager#listeners.

The listener won't be removed if we don't close the ReadOnlyLedgerHandle, so there will be lots of listeners in the heap memory.

The heap dump:

<img width="1572" alt="image" src="https://github.com/user-attachments/assets/8b4cd6d3-b7bf-4ef2-9c47-2e076bdbbba0">

(cherry picked from commit 999cd0f2ab14404be4d6c24e388456dbe56bb1a8)