Change the new ensemble log to info level (#4566)
---

### Motivation

Change the new ensemble log to the info level. Sometimes, the ensemble
may not satisfied with the placement policy. The log should be info
level but not a warn level. Because it will fix by the auto recovery
later, so this just a information from the ensemble choose.

(cherry picked from commit a12943df277e698a1621d33c245913365d031257)